+++ esbuild.config.cjs
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
+/* eslint-disable @typescript-eslint/no-require-imports */
 const esbuild = require('esbuild');
 const path = require('path');